--- src/components/controls/index.ts
@@ -1,2 +1,0 @@
-export * from "./Control";
-export * from "./settings";
