--- src/components/controls/settings/index.ts
@@ -1,3 +1,0 @@
-export * from "./theme-toggle";
-export * from "./FontToggle";
-export * from "./Settings";
