--- src/components/controls/settings/theme-toggle/index.ts
@@ -1,2 +1,0 @@
-export * from "./ThemeProvider";
-export * from "./ThemeToggle";
