--- src/components/index.ts
@@ -1,3 +1,0 @@
-export * from "./Button";
-
-export * from "./controls";
