--- src/libs/constants/index.ts
@@ -1,3 +1,0 @@
-export * from "./config";
-export * from "./url";
-export * from "./font-list";
