--- src/libs/enums/index.ts
@@ -1,1 +1,0 @@
-export * from "./toaster-type";
