--- src/libs/models/index.ts
@@ -1,3 +1,0 @@
-export * from "./config";
-export * from "./node-version";
-export * from "./toaster";
