+++ src/libs/utils/win-path.ts
@@ -1,5 +1,5 @@
+import { DEFAULT_REMOVE_FROM_PATH } from "@/libs/constants/config";
 import { invoke } from "@tauri-apps/api/core";
-import { DEFAULT_REMOVE_FROM_PATH } from "../constants";
 
 export const setToPath = (path: string) => {
   invoke("get_user_path").then((message: unknown) => {
