--- src/store/index.ts
@@ -1,3 +1,0 @@
-export * from "./config";
-export * from "./app";
-export * from "./node";
